+++ src/Components/DataviewItem/types.ts
@@ -1,0 +1,7 @@
+export interface Person {
+  name: string;
+  mass: string;
+  height: string;
+  birth_year: string;
+  hair_color: string;
+}
