+++ src/Components/DataviewList/types.ts
@@ -1,0 +1,5 @@
+import { Person } from '../DataviewItem/types';
+
+export interface DataviewListProps {
+  data: Person[];
+}
