+++ src/Components/ErrorBoundary/types.ts
@@ -1,0 +1,12 @@
+export interface ErrorBoundaryState {
+  hasError: boolean;
+  error?: Error;
+}
+
+export interface ErrorBoundaryProps {
+  children?: React.ReactNode;
+}
+
+export interface ErrorButtonProps {
+  error: Error | null;
+}
